Browser Studio: Share JSX insertion codemod
