+++ packages/browser-studio/package.json
@@ -32,6 +32,7 @@
 		"@remotion/studio-shared": "workspace:*",
 		"@rspack/browser": "2.1.1",
 		"fflate": "0.8.2",
+		"prettier": "catalog:",
 		"react-refresh": "0.18.0",
 		"remotion": "workspace:*"
 	},
